Move vault to controllers
